Refactor, nogui added
