--- ConsoleApp2/Properties/AssemblyInfo.cs
@@ -1,36 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-// Ogólne informacje o zestawie są kontrolowane poprzez następujący 
-// zestaw atrybutów. Zmień wartości tych atrybutów, aby zmodyfikować informacje
-// powiązane z zestawem.
-[assembly: AssemblyTitle("DifferentSLIAutoLoader")]
-[assembly: AssemblyDescription("Simple console application for loading DifferentSLIAuto unsigned driver.")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("DifferentSLIAutoLoader")]
-[assembly: AssemblyCopyright("Copyright ©  2019")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
-// Ustawienie elementu ComVisible na wartość false sprawia, że typy w tym zestawie są niewidoczne
-// dla składników COM. Jeśli potrzebny jest dostęp do typu w tym zestawie z
-// COM, ustaw wartość true dla atrybutu ComVisible tego typu.
-[assembly: ComVisible(false)]
-
-// Następujący identyfikator GUID jest identyfikatorem biblioteki typów w przypadku udostępnienia tego projektu w modelu COM
-[assembly: Guid("862eb500-474f-4fe4-8797-d399d71f46bb")]
-
-// Informacje o wersji zestawu zawierają następujące cztery wartości:
-//
-//      Wersja główna
-//      Wersja pomocnicza
-//      Numer kompilacji
-//      Rewizja
-//
-// Możesz określić wszystkie wartości lub użyć domyślnych numerów kompilacji i poprawki
-// przy użyciu symbolu „*”, tak jak pokazano poniżej:
-// [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
