--- DifferentSLIAutoLoader.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.106
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DifferentSLIAutoLoader", "ConsoleApp2\DifferentSLIAutoLoader.csproj", "{862EB500-474F-4FE4-8797-D399D71F46BB}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{862EB500-474F-4FE4-8797-D399D71F46BB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{862EB500-474F-4FE4-8797-D399D71F46BB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{862EB500-474F-4FE4-8797-D399D71F46BB}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{862EB500-474F-4FE4-8797-D399D71F46BB}.Debug|x86.Build.0 = Debug|Any CPU
-		{862EB500-474F-4FE4-8797-D399D71F46BB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{862EB500-474F-4FE4-8797-D399D71F46BB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{862EB500-474F-4FE4-8797-D399D71F46BB}.Release|x86.ActiveCfg = Release|Any CPU
-		{862EB500-474F-4FE4-8797-D399D71F46BB}.Release|x86.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {2E279637-AA21-4E9A-94F2-9E4E9A24A4DD}
-	EndGlobalSection
-EndGlobal
