+++ DifferentSLIAutoTools/Form1.cs
@@ -1,15 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
-using System.Management;
-using System.Threading;
-using System.Diagnostics;
 
 namespace DifferentSLIAutoTools
 {
@@ -27,39 +17,12 @@ namespace DifferentSLIAutoTools
 
         private void button1_Click(object sender, EventArgs e)
         {
-            Process s = new Process();
-            s.StartInfo.FileName = System.AppDomain.CurrentDomain.BaseDirectory + "shark\\Sea.exe";
-            s.StartInfo.Verb = "runas";
-            s.Start();
-            s.WaitForExit();
-            Thread.Sleep(1000);
-            Process d = new Process();
-            d.StartInfo.FileName = System.AppDomain.CurrentDomain.BaseDirectory + "dsefix\\dsefix.exe";
-            d.StartInfo.Verb = "runas";
-            d.Start();
-            d.WaitForExit();
-            ManagementObjectSearcher searcher =
-             new ManagementObjectSearcher("SELECT * FROM Win32_VideoController");
-            foreach (ManagementObject mo in searcher.Get())
-            {
-                PropertyData name = mo.Properties["Description"];
-                Process devManViewProc = new Process();
-                devManViewProc.StartInfo.FileName = System.AppDomain.CurrentDomain.BaseDirectory + "devmanview\\DevManView.exe";
-                devManViewProc.StartInfo.Arguments = "/disable_enable \"" + name.Value + "\"";
-                devManViewProc.StartInfo.Verb = "runas";
-                devManViewProc.Start();
-                devManViewProc.WaitForExit();
-            }
+            Logic.LoadDriver();
         }
 
         private void button2_Click(object sender, EventArgs e)
         {
-            Process de = new Process();
-            de.StartInfo.FileName = System.AppDomain.CurrentDomain.BaseDirectory + "dsefix\\dsefix.exe";
-            de.StartInfo.Arguments = "-e";
-            de.StartInfo.Verb = "runas";
-            de.Start();
-            de.WaitForExit();
+            Logic.dsefix(true);
         }
     }
 }
