+++ Install (Run as Admin).bat
@@ -1,4 +1,3 @@
 @echo off
-md "C:\DifferentSLIAutoLoader"
-xcopy %~dp0\DifferentSLIAutoLoader C:\DifferentSLIAutoLoader /s /y
-SchTasks /Create /SC ONSTART  /TN "DifferentSLIAutoLoader" /TR C:\DifferentSLIAutoLoader\DifferentSLIAutoLoader.exe /ru system /F
+SET path=%~dp0
+SchTasks /Create /SC ONSTART  /TN "DifferentSLIAutoLoader" /TR "%path%DifferentSLIAutoTools.exe /nogui" /ru system /F
