+++ Uninstall (Run as Admin).bat
@@ -1,3 +1,2 @@
 @echo off
-rmdir C:\DifferentSLIAutoLoader /q /s
 SchTasks /Delete /TN DifferentSLIAutoLoader /F
